requirements.txt
aiohttp
jsonschema
python-consul
hvac
prometheus_client
tarantool>=0.4